Initial commit, start with just book's file.
There are other files that come with osdev.tex which will be included
into this repository shortly. For now I'm starting with just the base
LaTeX book text as the other files are not unique to this project.

That is the other related files have code/text for other papers and
projects of mine. Sorry for the delay.